fix: match cookie max age to session max inactive interval
This fixes an issue where sessions are invalidated prematurely.

By default, sessions saved by Spring Session are only valid for
30 minutes (based on Tomcat server servlet defaults). This causes
an issue where cookies remain in the client browser, despite the
session it is referencing already being invalidated.

To fix this, we extend the session timeout to 7 days (the same as
the cookie max age) so that both the cookie and the session it is
referencing can be invalidated at the same time.

For more info, see:
[1] https://docs.spring.io/spring-boot/reference/web/spring-session.html
[2] https://www.codejava.net/frameworks/spring-boot/change-session-timeout